try to bundle only edge version of sentry
